content/canvas/test/test_2d.path.stroke.scale1.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.stroke.scale1</title>
<!-- Testing: Stroke line widths are scaled by the current transformation matrix -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
function isPixel(ctx, x,y, r,g,b,a, pos, colour, d) {
    var pixel = ctx.getImageData(x, y, 1, 1);
    var pr = pixel.data[0],
        pg = pixel.data[1],
        pb = pixel.data[2],
        pa = pixel.data[3];
    ok(r-d <= pr && pr <= r+d &&
       g-d <= pg && pg <= g+d &&
       b-d <= pb && pb <= b+d &&
       a-d <= pa && pa <= a+d,
       "pixel "+pos+" is "+pr+","+pg+","+pb+","+pa+"; expected "+colour+" +/- "+d);
}

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);

ctx.beginPath();
ctx.rect(25, 12.5, 50, 25);
ctx.save();
ctx.scale(50, 25);
ctx.strokeStyle = '#0f0';
ctx.stroke();
ctx.restore();

ctx.beginPath();
ctx.rect(-25, -12.5, 150, 75);
ctx.save();
ctx.scale(50, 25);
ctx.strokeStyle = '#f00';
ctx.stroke();
ctx.restore();

isPixel(ctx, 0,0, 0,255,0,255, "0,0", "0,255,0,255", 0);
isPixel(ctx, 50,0, 0,255,0,255, "50,0", "0,255,0,255", 0);
isPixel(ctx, 99,0, 0,255,0,255, "99,0", "0,255,0,255", 0);
isPixel(ctx, 0,25, 0,255,0,255, "0,25", "0,255,0,255", 0);
isPixel(ctx, 50,25, 0,255,0,255, "50,25", "0,255,0,255", 0);
isPixel(ctx, 99,25, 0,255,0,255, "99,25", "0,255,0,255", 0);
isPixel(ctx, 0,49, 0,255,0,255, "0,49", "0,255,0,255", 0);
isPixel(ctx, 50,49, 0,255,0,255, "50,49", "0,255,0,255", 0);
isPixel(ctx, 99,49, 0,255,0,255, "99,49", "0,255,0,255", 0);

SimpleTest.finish();

});
</script>
